src/third_party/mozjs/extract/intl/uconv/tests/test_utf8_overconsumption.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=445886
-->
<head>
  <meta http-equiv="Content-type" content="text/html; charset=UTF-8">
  <title>Test for Unicode non-characters</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript">
function Inject()
{
  $("display").innerHTML = "Evil";
}
  </script>
</head>
<body Â>onload="Inject()">
<pre id="test">
<script class="testbody" type="text/javascript">

/** test that UTF-8 decoding resynchronizes after incomplete sequences */
function test()
{
    is($("display").innerHTML, "All good.", "No overconsumption");
    SimpleTest.finish();
}

  addLoadEvent(function() {
    setTimeout(test, 0);
  });
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=445886">Mozilla Bug 445886</a>
<p id="display">All good.</p>
<div id="content" style="display: none"></div>
</body>
</html>
